Add a control panel and level system

+++ events/message.js
@@ -1,8 +1,57 @@
 const { WebhookClient } = require('discord.js');
 const ratelimit = new Set();
+const messageRatelimit = new Set();
+
+const mongoose = require('mongoose');
+var Int32 = require('mongoose-int32');
+const guildUserSchema = require.main.require('./schemes/guildUserSchema.js')
+const GuildUser = mongoose.model('guildUser', guildUserSchema, 'guildUser')
+
+async function createGuildUser(id, totalXP, level, xp, guildId) {
+  return new GuildUser({
+    id,
+    totalXP,
+    level,
+    xp,
+    guildId,
+  }).save()
+}
+
+async function findGuildUser(id, guildId) {
+  return await GuildUser.findOne({ id: id, guildId: guildId })
+}
 
 module.exports = async (client, message) => {
   if (message.author.bot) return;
+
+  if(message.guild && !messageRatelimit.has(message.guild.id  + "" + message.author.id)) {
+    let user = await client.connector.then(async () => {
+      return findGuildUser(message.author.id, message.guild.id)
+    })
+    
+    if (!user) {
+      user = await createGuildUser(message.author.id, 0, 0, 0, message.guild.id)
+    }
+
+    var xp = 15 + Math.round(Math.random() * 10);
+    var userXp = user.xp + xp;
+    user.totalXP = user.totalXP + xp;
+
+    if(userXp > 5 * (user.level ^ 2) + (50 * user.level) + 100) {
+      user.level++;
+      user.xp = userXp - (5 * (Math.pow(user.level,2)) + (50 * user.level) + 100);
+    } else {
+      user.xp = userXp;
+    }
+
+    user.save();
+
+    var timeout = client.config.messageLevelRatelimit;
+    messageRatelimit.add(message.guild.id  + ""  + message.author.id);
+    setTimeout(() => {
+      messageRatelimit.delete(message.guild.id  + ""  + message.author.id);
+    }, timeout)
+  }
 
   const settings = message.settings = client.getSettings(message.guild);
 
@@ -32,9 +81,11 @@ module.exports = async (client, message) => {
     message.react("🇦");
   }
 
-  if(message.channel.parent.id == "786153273186975765" && !message.member.roles.cache.has("785422260063698974")) {
-    message.delete();
-    return;
+  if(message.channel.parent) {
+    if(message.channel.parent.id == "786153273186975765" && !message.member.roles.cache.has("785422260063698974")) {
+      message.delete();
+      return;
+    }
   }
 
   // Also good practice to ignore any message that does not start with our prefix,
@@ -65,6 +116,10 @@ module.exports = async (client, message) => {
   // and return a friendly error message.
   if (cmd && !message.guild && cmd.conf.guildOnly)
     return message.channel.send("This command is unavailable via private message. Please run this command in a guild.");
+
+  if(cmd && !cmd.conf.enabled) {
+    return message.channel.send("This command is currently disabled!");
+  }
 
   if (level < client.levelCache[cmd.conf.permLevel]) {
     if (client.settings.systemNotice === "true") {
